Remove reference and timestamps are effectively remove_column migrations

+++ lib/safe_migrations/migration_ext.rb
@@ -27,6 +27,10 @@ module SafeMigrations
           case method
           when :remove_column
             raise UnsafeRemoveColumn
+          when :remove_timestamps
+            raise UnsafeRemoveColumn
+          when :remove_reference
+            raise UnsafeRemoveColumn
           when :change_table
             raise UnsafeChangeTable
           when :rename_table
